% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_bar.R
\name{ses_bar}
\alias{ses_bar}
\title{SES plot with a horizontal grid and without a border (useful for bar graphs)}
\usage{
ses_bar(legends = TRUE)
}
\arguments{
\item{legends}{If the legend needs to be displayed, the input should be TRUE.
If the legend is not needed, the input should be FALSE.}
}
\description{
A graph with a horizontal grid and no border is plotted.
This is useful for plotting a bar graph.
For more information, please visit \url{https://www.ses21.com}.

This is slightly different from ses_hist().
}
