% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_bar_theme.R
\name{ses_bar_theme}
\alias{ses_bar_theme}
\title{SES plot with a horizontal grid (a theme for bar graphs).}
\usage{
ses_bar_theme(legends = FALSE, borders = TRUE)
}
\arguments{
\item{legends}{If the legend needs to be displayed, the input should be TRUE.
If the legend is not needed, the input should be FALSE.}

\item{borders}{If the border needs to be displayed, the input should be TRUE.
If the border is not needed, the input should be FALSE.}
}
\description{
A graph with a horizontal grid is plotted. Border can be added or removed.
This is useful for plotting a bar graph.
For more information, please visit \url{https://www.ses21.com}.

This is slightly different from ses_hist().
}
