% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_bland_altman.R
\name{ses_bland_altman}
\alias{ses_bland_altman}
\title{SES themed Bland Altman plot}
\usage{
ses_bland_altman(statBlandAlt, point_size = 3.3, diff_ci = TRUE, ...)
}
\arguments{
\item{statBlandAlt}{Results from ses_statBlandAlt(), which is a function
that calculates important indices from drawing a Bland-Altman plot.}

\item{point_size}{The size of the individual points. The default is set to 3.3.}

\item{diff_ci}{If set TRUE, then it will draw a shaded region that represents the 95%
confidence interval of the difference between the two sessions from one-sample t-test.
If the region (i.e. confidence interval) overlaps with zero, then there
is no significant bias/difference between the two sessions/datasets.
If it does not overlap with 0, then the measurement variability is significantly large.}

\item{...}{Parameters of geom_point(), such as 'color', 'fill', 'shape', etc.}
}
\description{
This function generates a Bland-Altman plot using the output
from ses_statBlandAlt. The plot automatically uses ses_classic() theme.
The upper dashed line indicates the upper limit (mean_diff + 1.96*sd), the middle
dashed line indicates the mean difference between the
two samples, and the lower dashed line indicates the lower limit
(mean_diff - 1.96*sd).

To add a legend, you will need to add ses_classic(legends = TRUE).
}
\examples{
\dontrun{
set.seed(1)
first <- rnorm(20)
second <- rnorm(20)
df <- as_tibble(cbind(first,second)) # requires library(tidyverse)
res <- ses_statBlandAlt(df$first, df$second)
ggplot(data = res$data, aes(x = res$mean, y = res$diff)) +
ses_bland_altman(res)
}

}
