% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_boxplot.R
\name{ses_boxplot}
\alias{ses_boxplot}
\title{violin plot with individual points (jittered)}
\usage{
ses_boxplot(
  point_fill,
  boxplot_fill_color = "gray95",
  boxplot_width_color = "black",
  points = TRUE,
  notch = F,
  notchwidth = 0.4,
  point_size = 1,
  point_width_color = "white"
)
}
\arguments{
\item{point_fill}{Color of the points.}

\item{boxplot_fill_color}{Color of the boxplot.}

\item{boxplot_width_color}{Color of the boxplot's width.}

\item{points}{Logical argument is required. TRUE if points need to be displayed.
FALSE if points need to be not shown.}

\item{point_size}{Size of the points.}

\item{point_width_color}{Color of the point's width when shapes of the points are 21-25.}

\item{point_color}{Color of the points.}
}
\value{

}
\description{
A violin plot superimposed with jittered individual points.
}
\examples{

}
