% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_classic.R
\name{ses_classic}
\alias{ses_classic}
\title{SES classical theme.}
\usage{
ses_classic(legends = FALSE)
}
\arguments{
\item{legends}{If the legend needs to be displayed, the input should be TRUE.
If the legend is not needed, the input should be FALSE.}
}
\description{
It has x and y axis but no grids.
For more information, please visit \url{https://www.ses21.com}.
}
