% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_color.R
\name{ses_color}
\alias{ses_color}
\title{SES custom colors}
\usage{
ses_color(colorNum = 10)
}
\arguments{
\item{colorNum}{A numerical input that indicates the number of colors.}
}
\value{

}
\description{
This is a custom color palette that SES recommends for data visualization.
For more information, please visit \url{https://www.ses21.com}. It returns
up to 16 different colors with a high contrast.
}
\examples{
ses_color(2) # returns two colors

# plot a bar graph with a different color for each group

df <- data.frame(group=c("One", "Two", "Three"),
score=c(4.8, 9, 8.2))

p <- df \%>\%
ggplot(aes(x=group, y=score, fill = group)) +
geom_bar(stat='identity')

p +
ses_bar_border() +
ses_color(nrow(df)) # number of colors = number of groups
}
