% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_corr.R
\name{ses_corr}
\alias{ses_corr}
\title{SES plot without a border (useful for correlation plots)}
\usage{
ses_corr()
}
\arguments{
\item{...}{No input is required.}
}
\description{
A graph with no border is plotted. It has major vertical and horizontal grids. This is useful for plotting correlation.
For more information, please visit \url{https://www.ses21.com}.
}
