% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_corrline.R
\name{ses_corrline}
\alias{ses_corrline}
\title{Linear regression slope from correlation}
\usage{
ses_corrline(
  x,
  y,
  data = data,
  line_color = "black",
  line_size = 0.6,
  line_linetype = "dashed"
)
}
\arguments{
\item{x}{Numeric vectors of data (ex. group 1). This corresponds to data in the
x-axis of the correlation plot.

If the argument 'data' is also included, then x can be the column name
of 'data'.}

\item{y}{Numeric vectors of data (ex. group 2). This corresponds to data in the
y-axis of the correlation plot.

If the argument 'data' is also included, then y can be the column name
of 'data'.}

\item{line_color}{Color of the linear regression from the correlation.
The input should be a character string.}

\item{line_size}{Line width of the linear regression from the correlation.
It should be a number.}

\item{line_linetype}{Line type of the linear regression from the correlation.
The input should be a character string: 'solid', 'dashed', 'dotted',
'blank', 'dotdash', 'longdash', and 'twodash'.}
}
\value{

}
\description{
This function plots the linear regression of a correlation.
This does so by obtain coefficients from the lm() function using
two datasets (x and y).
It is recommended that the unity slope is plotted only when both
axes are scaled to each other (ex. z-scores).
}
\examples{

}
