% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_corrStat.R
\name{ses_corrStat}
\alias{ses_corrStat}
\title{Annotation of results from a paired correlation test on ggplot2}
\usage{
ses_corrStat(
  x,
  y,
  text_x,
  text_y,
  data = .data,
  method = "pearson",
  p = TRUE,
  r = TRUE,
  text_color = "black",
  font = "",
  text_size = 4,
  lin_reg = T,
  line_size = 1,
  line_type = "dashed",
  line_color = "black"
)
}
\arguments{
\item{x}{Data set of the x-axis in the correlation plot.
If the argument 'data' is filled, then x can be the column name.
Otherwise, it has to take a form of numerical vector.}

\item{y}{Data set of the y-axis in the correlation plot.
If the argument 'data' is filled, then y can be the column name.
Otherwise, it has to take a form of numerical vector.
x and y must have the same length.}

\item{text_x}{Location of the text in the x-axis of the plot.}

\item{text_y}{Location of the text in the y-axis of the plot.}

\item{data}{Dataframe that contains x and y. This argument is optional.}

\item{method}{Method of computing the correlation coefficient: 'pearson', 'kendall',
or 'spearman'.
'Pearson' is often used for normally distributed data.
'Spearman' is often used for computing the strength of a non-linear correlation.}

\item{p}{Logical input is required ('TRUE' for displaying the p-value, '
FALSE' for not displaying the p-value).
p-value is used for determining statistical significance in the
hypothesis test. It depends on the sample size.}

\item{r}{Logical input is required ('TRUE' for displaying the r-value, '
FALSE' for not displaying the r-value).
r-value is the correlation coefficient. It does not depend on}

\item{text_color}{Color of the text in character string (ex. 'black').}

\item{font}{Font of the text in character string.}

\item{size}{Font size of the text}
}
\value{

}
\description{
This function pastes the texts of the results from a correlation test
on ggplot2. For more information, please visit \url{https://www.ses21.com}.
}
\examples{

}
