% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_hgrid.R
\name{ses_hgrid}
\alias{ses_hgrid}
\title{SES theme with a horizontal grid}
\usage{
ses_hgrid(legends = FALSE, borders = TRUE)
}
\arguments{
\item{legends}{If the legend needs to be displayed, the input should be `TRUE`.
If the legend is not needed, the input should be `FALSE`.}

\item{borders}{If the border needs to be displayed, the input should be TRUE.
If the border is not needed, the input should be `FALSE`.}
}
\description{
A graph with a horizontal grid is plotted. Border can be added or removed.
This is useful for plotting a bar graph.
For more information, please visit \url{https://www.ses21.com}.

ses_bar_theme() is exactly the same as ses_hgrid().
}
