% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_hvgrid_minor.R
\name{ses_hvgrid_minor}
\alias{ses_hvgrid_minor}
\title{SES theme with horizontal and vertical major and minor grids}
\usage{
ses_hvgrid_minor(legends = TRUE, borders = TRUE)
}
\arguments{
\item{legends}{If the legend needs to be displayed, the input should be TRUE.
If the legend is not needed, the input should be FALSE.}

\item{borders}{If the border needs to be displayed, the input should be `TRUE`.
If the border is not needed, the input should be `FALSE`.}
}
\description{
This theme has vertical and horizontal grids.
It is useful for the lollipop plot.
For more examples, please visit \url{https://www.ses21.com/}.
}
