% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_minimal.R
\name{ses_minimal}
\alias{ses_minimal}
\title{SES minimal theme (no grid).}
\usage{
ses_minimal(legends = TRUE, borders = TRUE)
}
\arguments{
\item{legends}{If the legend needs to be displayed, the input should be `TRUE`.
If the legend is not needed, the input should be `FALSE`.}

\item{borders}{If the border needs to be displayed, the input should be `TRUE`.
If the border is not needed, the input should be `FALSE`}
}
\description{
This theme has no major grid.
For more information, please visit \url{https://www.ses21.com}.
}
