% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_palette.R
\name{ses_palette}
\alias{ses_palette}
\title{SES custom palette of colors
This is a custom color palette that SES recommends for data visualization.
For more information, please visit \url{https://www.ses21.com}. It returns
up to 16 different colors with a high contrast.}
\usage{
ses_palette(colorNum = 10)
}
\arguments{
\item{colorNum}{Number of colors (1-16).}
}
\value{

}
\description{
SES custom palette of colors
This is a custom color palette that SES recommends for data visualization.
For more information, please visit \url{https://www.ses21.com}. It returns
up to 16 different colors with a high contrast.
}
\examples{
ses_palette(3) # returns 3 colors

ggplot(data = mpg, mapping = aes(x = displ, y = hwy, color = class)) +
geom_point(size = 2) +
scale_color_manual(values = ses_palette(7)) +
ses_corr_theme()

}
