% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_slope.R
\name{ses_slope_theme}
\alias{ses_slope_theme}
\title{SES plot with a theme appropriate for the slope chart}
\usage{
ses_slope_theme(legends = TRUE)
}
\arguments{
\item{legends}{#' If the legend needs to be displayed, the input should be TRUE.
If the legend is not needed, the input should be FALSE.}
}
\description{
In this plot, all aspects except for the left-handed spine are missing.
This format is appropriate for the slope chart. For more information,
please visit \url{https://www.ses21.com}.
}
