% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_statBlandAlt.R
\name{ses_statBlandAlt}
\alias{ses_statBlandAlt}
\title{Statistics for a Bland-Altman plot}
\usage{
ses_statBlandAlt(first, second)
}
\arguments{
\item{first}{Data from the first repetition/session}

\item{second}{Data from the second repetition/session}
}
\description{
Bland-Altman plot is drawn to
show measurement variability/reliabiilty of a task. This
function requires two paired datasets (same length).
It returns a list of difference (by element), mean,
standard deviation of the difference,
mean difference, upper and lower limits. These values
are necessary to draw a Bland Altman plot.
The list returned from this function can be directly
used as an argument for ses_bland_altman(), which
draws a Bland-Altman plot using ggplot2.
}
