% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_add_line.R
\name{sm_add_line}
\alias{sm_add_line}
\title{Adding a line annotation in the combined plot}
\usage{
sm_add_line(
  combined_plot,
  x,
  y,
  xend,
  yend,
  color = "black",
  linewidth = 0.5,
  ...
)
}
\arguments{
\item{combined_plot}{Combined figure, an output from sm_put_together().}

\item{x}{Starting location of the line along the x-axis of the combined figure. The middle origin is at 0.5. Values from 0 to 1.}

\item{y}{Starting location of the line along the y-axis of the combined figure. The middle origin is at 0.5. Values from 0 to 1.}

\item{xend}{Final location of the line along the x-axis of the combined figure. The middle origin is at 0.5. Values from 0 to 1.}

\item{yend}{Final location of the line along the y-axis of the combined figure. The middle origin is at 0.5. Values from 0 to 1.}

\item{color}{Color of the line. Default is set to black.}

\item{linewidth}{Thickness of the line. Default is set to 0.5.}

\item{...}{Other parameters of the line that will be transferred to the function annotate()
from ggplot2.}
}
\value{
Prints a line to the combined plot.
}
\description{
Adding a line annotation in the combined plot
}
\examples{
\dontrun{
combined_figure2 <- sm_add_arrow(combined_figure, x = .5, y= .5, xend=.6, yend=.6)
}
}
