% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_bar.R
\name{sm_bar}
\alias{sm_bar}
\title{A bar plot with individual points (jittered)}
\usage{
sm_bar(
  bar_fill_color,
  bar_width = 0.7,
  bar_border_color = "transparent",
  bar_alpha = 1,
  errSize = 1,
  errColor = "black",
  point_size = 2.5,
  point_alpha = 0.65,
  se = T,
  point_jitter_width = 0.12,
  ...
)
}
\arguments{
\item{bar_fill_color}{Color of the all bars in the plot. It only accepts one color.
This argument is OPTIONAL.
If this argument is included, then all bars will have the same color.}

\item{bar_width}{Width of the bar.}

\item{bar_border_color}{Color of the bar borders. It only accepts one color.}

\item{bar_alpha}{Transparency of the bars (0 to 1).}

\item{errSize}{The line width of the error bar.}

\item{errColor}{The color of the error bar}

\item{point_size}{Size of the individual jittered points.}

\item{point_alpha}{Transparency of the jittered points.}

\item{se}{When the error bar should represent standard error,
it should be set to TRUE. If the error bar should
represent standard deviation, it should be set to FALSE.}

\item{point_jitter_width}{Width of the jitter}

\item{...}{Other parameters, such as 'color', 'shape', 'fill',
to specify the properties of the points.
For more information, type ?geom_point}
}
\description{
A bar plot superimposed by jittered individual points
and the error bar for standard error or standard deviation.
}
