% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_boxplot.R
\name{sm_boxplot}
\alias{sm_boxplot}
\title{A boxplot with individual points (jittered)}
\usage{
sm_boxplot(
  boxplot_fill_color = "gray95",
  boxplot_border_color = "black",
  points = TRUE,
  notch = F,
  width = 0.5,
  point_size = 2.5,
  ...
)
}
\arguments{
\item{boxplot_fill_color}{Color of the boxplot.}

\item{boxplot_border_color}{Color of the boxplot's border.}

\item{points}{#' TRUE if points need to be shown.
FALSE if points need to be hidden.}

\item{notch}{TRUE to notch the shape of the boxplot.
FALSE to keep the boxplot's shape.}

\item{width}{Width of the boxplot.}

\item{point_size}{Size of the individual jittered points.}

\item{...}{Other parameters for geom_point(), such as "fill", "color" and "shape".
For more information check out ?geom_point.}
}
\description{
A boxplot superimposed by jittered individual points.
}
