% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_color.R
\name{sm_color}
\alias{sm_color}
\title{SM custom palette of colors}
\usage{
sm_color(...)
}
\arguments{
\item{...}{The input has to be a character string.of a color name.
There are 16 colors available from the SES palette:
`'blue'`,`'cyan'`, `'green'`, `'purple'`, `'orange'`,
`'skyblue'`, `'pink'`, `'limegreen'`, `'lightpurple'`,
`'brown'`, `'red'`, `'lightorange'`,
`'asparagus'`, `'viridian'`, `'darkred'`, `'lightblue'`}
}
\value{
A character/string of hex codes
}
\description{
This is a custom color palette that SES recommends for data visualization.
It returns up to 16 different colors with a high visibility.
}
\examples{
sm_color('cyan')

sm_color('cyan', 'green', 'blue')
}
