% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_common_axis.R
\name{sm_common_axis}
\alias{sm_common_axis}
\title{A function to plot panels with common x- and y- axes}
\usage{
sm_common_axis(location, hmargin = 1, wmargin = 1)
}
\arguments{
\item{location}{Location of the panel.
`'topleft'`: removes x-axis title, x-axis ticklabel, y-axis title.
`'topright'`:  removes x-axis title, x-axis ticklabel, y-axis title, y-axis ticklabel.
`'bottomleft'`: removes x-axis title, y-axis title.
`'bottomright'`: removes x-axis title, y-axis title, y-axis ticklabel.
`'topcenter'`: removes x-axis title, x-axis ticklabel, y-axis title, y-axis ticklabel.
`'bottomcenter'`: removes x-axis title, y-axis title, y-axis ticklabel.}

\item{hmargin}{The amount of height of blank space between subplots. It sets the size of the empty space (i.e., margin) between panels. T
he default is set to 1, which should reduce the empty space (right and left side of each panel)
between the panels.}

\item{wmargin}{The amount of width of blank space between subplots. It sets the size of the empty space (i.e., margin) between panels. T
he default is set to 1, which should reduce the empty space (right and left side of each panel)
between the panels.}
}
\description{
This function should be used when plotting individual data.
Please read Chapter 5 of the online guide for more
details: smin95.github.io/dataviz
}
