% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_common_title.R
\name{sm_common_title}
\alias{sm_common_title}
\title{Common title for combined subplots}
\usage{
sm_common_title(title = "", size = 17, x = 0.5, y = 0.5, fontface = "bold")
}
\arguments{
\item{title}{The input should be string.}

\item{size}{Text size of the title.}

\item{x}{Location of the title along the x-axis. Default is the middle origin (0.5).}

\item{y}{Location of the legend along the y-axis. Default is the middle origin (0.5).}

\item{fontface}{The default is to set the text of the title as bold. This can be changed, to
either "plain", "bold", "italic", "bold.italic" .}
}
\value{
It prints title on a blank layer of ploting.
}
\description{
Common title for combined subplots
}
\examples{
library(smplot2)
sm_common_title('My title')

}
