% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_corr_theme.R
\name{sm_corr_theme}
\alias{sm_corr_theme}
\title{SM theme for correlation plots}
\usage{
sm_corr_theme(...)
}
\arguments{
\item{...}{This is the parameter for `sm_slope_all`.}
}
\description{
This theme has major vertical and horizontal grids. This is useful for plotting correlation.
sm_corr_theme() is exactly the same as sm_hvgrid(). For more information please
look up `?sm_hvgrid`.
}
\examples{
\dontrun{
library(tidyverse)
ggplot(data = mpg) +
geom_point(mapping = aes(x = displ, y = hwy, color = class)) +
sm_corr_theme()
}

}
