% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_forest.R
\name{sm_forest}
\alias{sm_forest}
\title{A forest plot - a combination of average point, error bar and individual points}
\usage{
sm_forest(
  data,
  x,
  y,
  range_side = "top",
  errorbar_type = "ci",
  sep_level = 2,
  vertical = FALSE,
  jitter_width = 0,
  point_size = 2.5,
  avg_point_shape = 18,
  err_color = "black",
  avg_point_size = 5,
  err_text_size = 4,
  borders = TRUE,
  legends = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame. Each column has to be a variable.
Each row has to be an observation. This data will get used as
argument in ggplot().}

\item{x}{Column variable of the data that defines the level of the x-axis.
This is used to define the aes() in ggplot().}

\item{y}{Column variable of the data that defines the level of the y-axis.
This is used to define the aes() in ggplot().}

\item{range_side}{String argument to specify the side of the text of the error range (ex. 95% CI interval texts).
The options are: 'top' and 'bottom'. This is an similar argument to
'which_side' in the `sm_raincloud()` function.}

\item{errorbar_type}{This argument determines the errorbar type.
If it is set to 'se' , standard error bar will be shown.
If it is set to 'sd', the error bar will display standard deviation.
If it is set to 'ci' (default), the error bar will display 95\% confidence interval.}

\item{sep_level}{A numerical value (0-3) that controls the level of the separation among
the individual points and the average point. The value can be 0-4.
If it's 0, all of these are clustered together. If it's 3, they are all
separated. 1 and 2 are somewhere in the middle. Default is set to 2.}

\item{vertical}{The orientation of the plots. The default is set to FALSE.
If you want the horizontal orientation of the plot, please set this argument
as FALSE.}

\item{jitter_width}{A numerical value that determines the degree of the jitter for each point. If its 0,
all the points will have no jitter (aligned along the y-axis).}

\item{point_size}{Size of the individual points.}

\item{avg_point_shape}{Shape of the average point.}

\item{err_color}{Color of the error bar.}

\item{avg_point_size}{Size of the avrage point}

\item{err_text_size}{Size of the texts that indicate the width of the error.}

\item{borders}{If the border needs to be displayed, the input should be TRUE.
If the border is not needed, the input should be FALSE.}

\item{legends}{If the legend needs to be displayed, the input should be TRUE.
If the legend is not needed, the input should be FALSE.}

\item{...}{Parameters for the points' aesthetics, such as 'fill', 'shape', 'color',
and 'alpha'. For more information, please type ?geom_point}
}
\description{
This function visualizes a forest plot, which is most commonly used when describing
effect size across various studies in meta-analyses.
}
\examples{
\dontrun{
#' set.seed(2) # generate random data
day1 = rnorm(20,0,1)
day2 = rnorm(20,5,1)
day3 = rnorm(20,6,1.5)
day4 = rnorm(20,7,2)
Subject <- rep(paste0('S',seq(1:20)), 4)
Data <- data.frame(Value = matrix(c(day1,day2,day3,day4),ncol=1))
Day <- rep(c('Day 1', 'Day 2', 'Day 3', 'Day 4'), each = length(day1))
df2 <- cbind(Subject, Data, Day)

sm_forest(data = df2, x = Value, y = Day, alpha = 0.2,
error_type = 'sd',
range_side = 'top',
jitter_width = 0.1, sep_level = 0,
avg_point_shape= 23, shape = 21)
}

}
