% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_palette.R
\name{sm_palette}
\alias{sm_palette}
\title{SM custom palette of colors}
\usage{
sm_palette(colorNum = 10)
}
\arguments{
\item{colorNum}{Number of colors (1-16).}
}
\description{
This is a custom color palette that SM recommends for data visualization.
It returns up to 16 different colors with a high contrast.
}
\examples{
sm_palette(3) # returns 3 colors

}
