% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_panel_sticker.R
\name{sm_panel_sticker}
\alias{sm_panel_sticker}
\title{Adding a sticker to each panel of a combined figure}
\usage{
sm_panel_sticker(all_plots, x, y, shape = 17, colors, sticker_size = 8, ...)
}
\arguments{
\item{all_plots}{all_plots should be list, which should contain all panels
that are to be combined into one figure.}

\item{x}{Location of the label along the x-axis (0 to 1). 0.5 is the middle origin.}

\item{y}{Location of the label along the y-axis (0 to 1). 0.5 is the middle origin.}

\item{shape}{Shape of the sticker}

\item{colors}{Colors of the stickers. It should be a vector of multiple colors. If only one is
provided, then the stickers on all panels will have the same color.}

\item{sticker_size}{Size of the sticker.}

\item{...}{Additional parameters for adjusting the appearance of the sticker. Same parameters
for annotate() from ggplot2.}
}
\value{
It returns a list of plots with panel stickers.
}
\description{
Adding a sticker to each panel of a combined figure
}
\examples{
\dontrun{
all_plots_sticker <- sm_panel_sticker(all_plots, x = 0.1, y = 0.9, colors='gray80')
}
}
