% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_put_together.R
\name{sm_put_together}
\alias{sm_put_together}
\title{Combining figures together}
\usage{
sm_put_together(
  all_plots,
  title,
  xlabel,
  ylabel,
  legend,
  ncol,
  nrow,
  tickRatio = 1.4,
  panel_scale = 0.9,
  wRatio = 1.15,
  hRatio = 1.11,
  hmargin = 0,
  wmargin = 0,
  remove_ticks = TRUE
)
}
\arguments{
\item{all_plots}{all_plots should be list, which should contain all panels
that are to be combined into one figure.}

\item{title}{Title layer that will determine the main title of the combined plot.
This is created using sm_common_title(). Optional argument.}

\item{xlabel}{xlabel layer that will determine the label of the combined plot's x-axis.
This is created using sm_common_xlabel(). Optional argument.}

\item{ylabel}{ylabel layer that will determine the label of the combined plot's y-axis.
This is created using sm_common_ylabel(). Optional argument.}

\item{legend}{ggplot() layer that has legend. Optional argument.}

\item{ncol}{Number of columns in the combined plot}

\item{nrow}{Number of rows in the combined plot}

\item{tickRatio}{Relative size of the ticks to the default aesthetics of the thematic functions (ex. sm_hgrid()).
If there are more rows or columns, please increase the tickRatio (1.4 - 1.8).}

\item{panel_scale}{Scale of the panel. Default is set to 0.9 to reduce empty space
within and around each panel. The user can set to a value from 0 to 1 to
see what happens to the spacing within each panel and between panels.}

\item{wRatio}{This adjusts the ratio of the width of the first column to those of other columns.
By default, it is set to be 1.15x wider than that of other columns. The 0.1 difference
is to compensate for the loss of space due to y-ticks in the first column.}

\item{hRatio}{This adjusts the ratio of the height of the last row to those of other rows
By default, it is set to be 1.11x taller than that of other columns. The 0.1 difference
is to compensate for the loss of space due to x-ticks in the last row.}

\item{hmargin}{The amount of height of blank space between subplots. It sets the size of the empty space (i.e., margin) between panels. T
he default is set to 0. If its positive, the blank spacing will increase. If its negative, it will get reduced
between panels.}

\item{wmargin}{The amount of width of blank space between subplots. It sets the size of the empty space (i.e., margin) between panels. T
he default is set to 0. If its positive, the blank spacing will increase. If its negative, it will get reduced
between panels.}

\item{remove_ticks}{X-axis ticks and y-axis ticks will be removed in inner plots.}
}
\value{
Returns a combined figure.
}
\description{
Combining figures together
}
\examples{
\dontrun{
title <- sm_common_title('My title')
xlabel <- sm_common_xlabel('My x-axis ')
ylabel <- sm_common_ylabel('My y-axis')
plots_tgd <- sm_put_together(indv_plots, title, xlabel,ylabel,
ncol=3,nrow=3)
}
}
