% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_raincloud.R
\name{sm_raincloud}
\alias{sm_raincloud}
\title{A raincloud plot - a combination of jittered points, boxplots
and violin plots.}
\usage{
sm_raincloud(
  data,
  x,
  y,
  which_side = "right",
  sep_level = 2,
  vertical = TRUE,
  jitter_width = 0.09,
  point_size = 3,
  violin_alpha = 0.3,
  boxplot_alpha = 1,
  borders = TRUE,
  legends = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame. Each column has to be a variable.
Each row has to be an observation. This data will get used as
argument in ggplot().}

\item{x}{Column variable of the data that defines the level of the x-axis.
This is used to define the aes() in ggplot().}

\item{y}{Column variable of the data that defines the level of the y-axis.
This is used to define the aes() in ggplot().}

\item{which_side}{String argument to specify the side of the boxplots and violinplots.
The options are: 'right', 'left', and 'mixed'. 'mixed' only works
when there are 2 levels in the x-axis. Otherwise, it will return an error.}

\item{sep_level}{A numerical value that controls the level of the separation among
the boxplot, violin plot and the points. The value can be 0-4.
If it's 0, all of these are clustered together. If it's 3, they are all
separated. 1 and 2 are somewhere in the middle. Default is set to 2.}

\item{vertical}{The orientation of the plots. The default is set to TRUE.
If you want the horizontal orientation of the plot, please set this argument
as FALSE.}

\item{jitter_width}{A numerical value that determines the degree of the jitter for each point. If its 0,
all the points will have no jitter (aligned along the y-axis).}

\item{point_size}{Size of the points.}

\item{violin_alpha}{Transparency of the violin (0 to 1).}

\item{boxplot_alpha}{Transparency of the boxplot (0 to 1).}

\item{borders}{If the border needs to be displayed, the input should be TRUE.
If the border is not needed, the input should be FALSE.}

\item{legends}{If the legend needs to be displayed, the input should be TRUE.
If the legend is not needed, the input should be FALSE.}

\item{...}{Parameters for the points' aesthetics, such as 'fill', 'shape', 'color',
and 'alpha'. For more information, please type ?geom_point}
}
\description{
This function visualizes a raincloud plot. The creation of this package
has been inspired by the R package called 'raincloudplots' by Jordy van
Langen (https://github.com/jorvlan/raincloudplots).

This function has been created to allow more customisation than the functions
in the raincloudplots package. Also, this function automatically sorts the data given the
condition that the x-axis factor levels have been sorted properly.
}
\examples{
\dontrun{
set.seed(2) # generate random data
day1 = rnorm(20,0,1)
day2 = rnorm(20,5,1)
day3 = rnorm(20,6,1.5)
day4 = rnorm(20,7,2)
Subject <- rep(paste0('S',seq(1:20)), 4)
Data <- data.frame(Value = matrix(c(day1,day2,day3,day4),ncol=1))
Day <- rep(c('Day 1', 'Day 2', 'Day 3', 'Day 4'), each = length(day1))
df2 <- cbind(Subject, Data, Day)

# use the random data to generate a raincloud plot
sm_raincloud(data = df2, x = Day, y = Value)
}
}
\references{
Allen M, Poggiali D, Whitaker K et al. Raincloud plots: a multi-platform tool for robust data visualization [version 2; peer review: 2 approved]. Wellcome Open Res 2021, 4:63. DOI: 10.12688/wellcomeopenres.15191.2
}
