% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_slope.R
\name{sm_slope}
\alias{sm_slope}
\title{A slope chart}
\usage{
sm_slope(
  labels,
  line_color = "gray53",
  line_size = 0.4,
  point_size = 2.5,
  line_type = "solid",
  line_alpha = 1,
  xtick_high = TRUE,
  legends = FALSE,
  ...
)
}
\arguments{
\item{labels}{Labels for the ticks of the x-axis. This is a required argument.
It has to be a single vector containing either one
or multiple elements. ex: c('Day 1', 'Day 2')}

\item{line_color}{Color of the lines}

\item{line_size}{Size of the lines}

\item{point_size}{Size of the points}

\item{line_type}{Type of the line. Default is 'solid'; other options include:
'twodash', 'longdash', 'dotted', 'dotdash', 'dashed', and
'blank'.}

\item{line_alpha}{Transparency of the lines (0 to 1).}

\item{xtick_high}{If it is set to TRUE, the ticks of the x-axis will be above. The argument
for 'labels' is required.
If it is set to FALSE, they will be at the bottom. Also the argument
for 'labels' will be ignored in this case. To change the labels, you
will need to add a geom object with scale_x_discrete(labels = c('A','B')).}

\item{legends}{If the legend needs to be displayed, the input should be TRUE.
If the legend is not needed, the input should be FALSE.}

\item{...}{Other parameters of geom_point(), such as 'shape', fill', and 'color.
For more information, please type ?geom_point.}
}
\description{
This function generates a slope chart.
This is very useful for comparing the effect between two time points.

For this function to work properly,
ggplot()'s mapping has to be quite specific.
1. Each observation has to be grouped.
}
\examples{
\dontrun{
set.seed(1) # generate random data
day1 = rnorm(20,0,1)
day2 = rnorm(20,5,1)
Subject <- rep(paste0('S',seq(1:20)), 2)
Data <- data.frame(Value = matrix(c(day1,day2),ncol=1))
Day <- rep(c('Day 1', 'Day 2'), each = length(day1))
df <- cbind(Subject, Data, Day) # final dataframe

#use the dataframe to generate a bar plot
ggplot(data = df, mapping = aes(x = Day, y = Value, group = Subject)) +
sm_slope(labels = c('Day 1', 'Day 2'))
}
}
