% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_slope_theme.R
\name{sm_slope_theme}
\alias{sm_slope_theme}
\title{SM plot with a theme appropriate for the slope chart}
\usage{
sm_slope_theme(legends = TRUE)
}
\arguments{
\item{legends}{#' If the legend needs to be displayed, the input should be TRUE.
If the legend is not needed, the input should be FALSE.}
}
\value{
Returns a background theme that is suitable for a slope chart (ggplot2 output).
}
\description{
In this plot, all aspects except for the left-handed spine are missing.
This format is appropriate for the slope chart.
}
\examples{
library(ggplot2)
library(smplot2)

ggplot(data = mpg) +
geom_point(mapping = aes(x = displ, y = hwy, color = class)) +
sm_slope_theme()

}
