% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_vgrid.R
\name{sm_vgrid}
\alias{sm_vgrid}
\title{SM theme with vertical grid}
\usage{
sm_vgrid(legends = TRUE, borders = TRUE)
}
\arguments{
\item{legends}{If the legend needs to be displayed, the input should be TRUE.
If the legend is not needed, the input should be FALSE.}

\item{borders}{If the border needs to be displayed, the input should be TRUE.
If the border is not needed, the input should be FALSE.}
}
\description{
This theme has major vertical grids.
}
