% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_bar.R
\name{ses_bar}
\alias{ses_bar}
\title{A SES bar plot with individual points (jittered)}
\usage{
ses_bar(
  data = data,
  x,
  y,
  bar_fill_color = "gray85",
  width = 0.4,
  point_size = 2.5,
  point_border_color = "white",
  point_shape = 21,
  ...
)
}
\arguments{
\item{data}{Entire dataset for plotting the jittered points.
Each row has to be each subject.
Column has to be group/variable/condition.}

\item{x}{Variable that defines x-axis. It should be independent variable.
Ideally, it should be identical to the x-axis mapped in aes() from ggplot().}

\item{y}{Variable that defines y-axis. It should be dependent variable.
Ideally, it should be identical to the y-axis mapped in aes() from ggplot().}

\item{bar_fill_color}{Color of the bar.}

\item{width}{Width of the bar.}

\item{point_size}{Size of the individual jittered points.}

\item{point_border_color}{Color of the points' border.}

\item{point_shape}{Shape of the jittered points.
Only shapes (21-25) with borders are allowed.}

\item{...}{Other parameters  to specify the properties of the points.
For more information, type ?geom_point}
}
\description{
A bar plot superimposed by jittered individual points.
ggplot() should contain the summarised data with
the mean for each group/condition.
}
