% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_bar.R
\name{ses_bar}
\alias{ses_bar}
\title{A SES bar plot with individual points (jittered)}
\usage{
ses_bar(
  data = data,
  x,
  y,
  bar_fill_color = "gray85",
  width = 0.4,
  point_size = 2.5,
  point_border_color = "white",
  point_shape = 21,
  ...
)
}
\arguments{
\item{data}{Entire dataset for plotting the jittered points.
Each row has to be each subject.
Column has to be group/variable/condition.}

\item{x}{Variable that defines x-axis. It should be independent variable.
Ideally, it should be identical to the x-axis mapped in aes() from ggplot().}

\item{y}{Variable that defines y-axis. It should be dependent variable.
Ideally, it should be identical to the y-axis mapped in aes() from ggplot().}

\item{bar_fill_color}{Color of the bar.}

\item{width}{Width of the bar.}

\item{point_size}{Size of the individual jittered points.}

\item{point_border_color}{Color of the points' border.}

\item{point_shape}{Shape of the jittered points.
Only shapes (21-25) with borders are allowed.}

\item{...}{}
}
\description{
A bar plot superimposed by jittered individual points.
ggplot() should contain the summarised data with
the mean for each group/condition.
}
\examples{
library(tidyverse)
library(sesplot)
set.seed(2) # generate random data
day1 = rnorm(20,0,2)
day2 = rnorm(20,6,2)
Subject <- rep(paste0('S',seq(1:20)), 2)
Data <- data.frame(Value = matrix(c(day1,day2),ncol=1))
Time <- rep(c('Day 1', 'Day 2'), each = length(day1))
DataFrame <- cbind(Subject, Data, Time)

# summarise the data with dplyr
summary_df <- DataFrame \%>\% group_by(Time) \%>\%
summarise(mean_value = mean(Value), se = se(Value))

# plot the data with ses_bar()
ggplot(data = summary_df, aes(x = Time, y = mean_value,
                             fill = Time)) +
ses_bar(data = DataFrame, x =
Time, y = Value) + ses_bar_theme() +
geom_linerange(aes(ymin = mean_value-se,
ymax = mean_value+se)) # errorbar
}
