% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_bar.R
\name{ses_bar}
\alias{ses_bar}
\title{A SES bar plot with individual points (jittered)}
\usage{
ses_bar(
  point_fill,
  data = .data,
  aes_x,
  aes_y,
  bar_fill_color = "gray85",
  width = 0.4,
  point_size = 3,
  point_width_color = "white",
  point_shape = 21
)
}
\arguments{
\item{point_fill}{Color of the individual jittered points.}

\item{data}{Entire dataset for plotting the jittered points.
Row has to be subject.
Column has to be group/variable/condition.}

\item{width}{Width of the bar}

\item{point_size}{Size of the individual jittered points.}

\item{point_width_color}{Color of the points' width.}

\item{point_shape}{Shape of the jittered points.}

\item{barplot_fill_color}{Color of the bar}

\item{points}{'TRUE' if points need to be shown.
'FALSE' if points need to be hidden.}
}
\value{

}
\description{
A bar plot superimposed with jittered individual points.
}
\examples{

}
