% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_bar_border.R
\name{ses_bar_border}
\alias{ses_bar_border}
\title{SES plot with a border (useful for bar plots)}
\usage{
ses_bar_border()
}
\arguments{
\item{...}{No input is required.}
}
\description{
A graph with a border is plotted. It is aesthetically appropriate for bar graphs.

For more information, please visit \url{https://www.ses21.com}.
}
