% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_bland_altman.R
\name{ses_bland_altman}
\alias{ses_bland_altman}
\title{SES themed Bland Altman plot}
\usage{
ses_bland_altman(statBlandAlt, point_size = 3.3, ...)
}
\arguments{
\item{statBlandAlt}{Results from ses_statBlandAlt(), which is a function
that calculates important indices from drawing a Bland-Altman plot.}

\item{point_size}{The size of the individual points. The default is set to 3.3.}

\item{...}{Parameters of geom_point(), such as 'color', 'fill', 'shape', etc.}
}
\description{
It automatically uses ses_classic() theme.
}
