% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_boxplot.R
\name{ses_boxplot}
\alias{ses_boxplot}
\title{A boxplot with individual points (jittered)}
\usage{
ses_boxplot(
  point_fill,
  boxplot_fill_color = "gray95",
  boxplot_width_color = "black",
  points = TRUE,
  notch = F,
  width = 0.4,
  point_size = 2,
  point_width_color = "white",
  point_shape = 21
)
}
\arguments{
\item{point_fill}{Color of the individual jittered points.}

\item{boxplot_fill_color}{Color of the boxplot.}

\item{boxplot_width_color}{Color of the boxplot's width.}

\item{points}{'TRUE' if points need to be shown.
'FALSE' if points need to be hidden.}

\item{notch}{'TRUE' to notch the shape of the boxplot.
'FALSE' to keep the boxplot's shape.}

\item{width}{Width of the boxplot.}

\item{point_size}{Size of the individual jittered points.}

\item{point_width_color}{Color of the points' width.}

\item{point_shape}{Shape of the jittered points.}
}
\value{

}
\description{
A boxplot superimposed with jittered individual points.
}
\examples{

}
