% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_boxplot.R
\name{ses_boxplot}
\alias{ses_boxplot}
\title{A boxplot with individual points (jittered)}
\usage{
ses_boxplot(
  boxplot_fill_color = "gray95",
  boxplot_border_color = "black",
  points = TRUE,
  notch = F,
  width = 0.5,
  point_size = 2.5,
  point_alpha = 0.35,
  point_border_color = "white",
  point_shape = 21,
  ...
)
}
\arguments{
\item{boxplot_fill_color}{Color of the boxplot.}

\item{boxplot_border_color}{Color of the boxplot's border.}

\item{points}{#' TRUE if points need to be shown.
FALSE if points need to be hidden.}

\item{notch}{TRUE to notch the shape of the boxplot.
FALSE to keep the boxplot's shape.}

\item{width}{Width of the boxplot.}

\item{point_size}{Size of the individual jittered points.}

\item{point_alpha}{Transparency of the jittered points.
This argument is ignored when points = FALSE.}

\item{point_border_color}{Color of the points' border.}

\item{point_shape}{Shape of the jittered points.
Only shapes (21-25) with borders are allowed.}

\item{...}{Other parameters for geom_point(), such as "fill".
For more information check out ?geom_point.}
}
\description{
A boxplot superimposed by jittered individual points.
}
