% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_color.R
\name{ses_color}
\alias{ses_color}
\title{SES custom palette of colors}
\usage{
ses_color(...)
}
\arguments{
\item{color}{The input has to be a character string.
There are 16 colors available from the SES palette:
`'blue'`,`'cyan'`, `'green'`, `'purple'`, `'orange'`,
`'lightblue'`, `'pink'`, `'limegreen'`, `'lightpurple'`,
`'brown'`, `'red'`, `'lightorange'`,
`'asparagus'`, `'viridian'`, `'darkred'`}
}
\value{

}
\description{
This is a custom color palette that SES recommends for data visualization.
For more information, please visit \url{https://www.ses21.com}. It returns
up to 16 different colors with a high visibility.
}
\examples{
ses_color('red')

ses_color('red', 'green', 'blue')
}
