% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_corrtext.R
\name{ses_corrtext}
\alias{ses_corrtext}
\title{Annotation of results from a correlation test on ggplot2}
\usage{
ses_corrtext(
  df1,
  df2,
  x,
  y,
  method = "pearson",
  p = TRUE,
  r = TRUE,
  text_color = "black",
  font = ""
)
}
\arguments{
\item{df1}{Numeric vectors of the first set of data (ex. group 1).}

\item{df2}{Numeric vectors of the second set of data (ex. group 2).
df1 and df2 must have the same length.}

\item{x}{Location of the text in the x-axis of the plot.}

\item{y}{Location of the text in the y-axis of the plot.}

\item{method}{Method of computing the correlation coefficient: 'pearson', 'kendall',
or 'spearman'.
'Pearson' is often used for normally distributed data.
'Spearman' is often used for computing the strength of a non-linear correlation.}

\item{p}{Logical input is required ('TRUE' for displaying the p-value, '
FALSE' for not displaying the p-value).
p-value is used for determining statistical significance in the
hypothesis test. It depends on the sample size.}

\item{r}{Logical input is required ('TRUE' for displaying the r-value, '
FALSE' for not displaying the r-value).
r-value is the correlation coefficient. It does not depend on}

\item{text_color}{Color of the text in character string (ex. 'black').}

\item{font}{Font of the text in character string.}
}
\value{

}
\description{
This function pastes the texts of the results from a correlation test
on ggplot2.
}
\examples{

}
