% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_hist.R
\name{ses_hist}
\alias{ses_hist}
\title{SES plot with a horizontal grid and without a border (useful for histogram)}
\usage{
ses_hist()
}
\arguments{
\item{...}{No input is required.}
}
\description{
A graph with a horizontal grid and no border is plotted. This is useful for plotting a histogram.
For more information, please visit \url{https://www.ses21.com}.
}
