% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_hvgrid.R
\name{ses_hvgrid}
\alias{ses_hvgrid}
\title{SES theme with horizontal and vertical grids}
\usage{
ses_hvgrid(legends = TRUE, borders = TRUE)
}
\arguments{
\item{legends}{If the legend needs to be displayed, the input should be TRUE.
If the legend is not needed, the input should be FALSE.}

\item{borders}{If the border needs to be displayed, the input should be TRUE.
If the border is not needed, the input should be FALSE.}
}
\description{
This theme has major vertical and horizontal grids. This is useful for plotting correlation.
ses_corr_theme() is exactly the same as ses_hvgrid().
For more information, please visit \url{https://www.ses21.com}.
}
