% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_slope.R
\name{ses_slope}
\alias{ses_slope}
\title{A SES slope chart}
\usage{
ses_slope(
  labels,
  line_color = "gray53",
  line_size = 0.4,
  point_size = 2.5,
  line_type = "solid",
  ...
)
}
\arguments{
\item{labels}{Labels for the ticks of the x-axis. This is a required argument.
It has to be a single vector containing either one
or multiple elements. ex: c('Day 1', 'Day 2')}

\item{line_color}{Color of the lines}

\item{line_size}{Size of the lines}

\item{point_size}{Size of the points}

\item{line_type}{Type of the line. Default is 'solid'; other options include:
'twodash', 'longdash', 'dotted', 'dotdash', 'dashed', and
'blank'.}

\item{...}{Other parameters of geom_point(), such as 'shape', fill', and 'color.
For more information, please type ?geom_point.}
}
\description{
This function generates a slope chart.
This is very useful for comparing the effect between two time points.

For this function to work properly,
ggplot()'s mapping has to be quite specific.
1. Each observation has to be grouped.
}
