% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_slope.R
\name{ses_slope}
\alias{ses_slope}
\title{A SES slope chart}
\usage{
ses_slope(
  line_color = "gray53",
  line_size = 0.4,
  point_border_color = "white",
  point_size = 2.5,
  point_shape = 21,
  point_stroke = 0.5,
  ...
)
}
\arguments{
\item{line_color}{Color of the lines}

\item{line_size}{Size of the lines}

\item{point_border_color}{Color of the point's border}

\item{point_size}{Size of the points}

\item{point_shape}{Shape of the points (21-25)}

\item{point_stroke}{Border width of the points}

\item{...}{Other parameters of scale_x_discrete. For more information,
please type ?scale_x_discrete.}
}
\description{
This function generates a slope chart.
This is very useful for comparing the effect between two time points.

For this function to work properly,
ggplot()'s mapping has to be quite specific.
1. Each observation has to be grouped.
2. In aes(), "fill" argument has to be filled in order to generate any filled color.
}
