% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_statCorr.R
\name{ses_statCorr}
\alias{ses_statCorr}
\title{Title
Linear regression slope and statistical values
from a paired correlation test}
\usage{
ses_statCorr(
  ...,
  line_type = "dashed",
  corr_method = "pearson",
  ln_method = "lm",
  se = F,
  line_size = 1,
  separate_by = ",",
  label_x = NULL,
  label_y = NULL,
  text_size = 4
)
}
\arguments{
\item{...}{Arguments for the properties of regression line.
For more information, type ?geom_smooth}

\item{line_type}{Options include:  'blank', 'solid', 'dashed', 'dotted',
'dotdash', 'longdash', 'twodash'.}

\item{corr_method}{Method of the correlation test.
Options include: 'pearson', 'kendall', or 'spearman'.}

\item{ln_method}{This is the method of producing the best-fit line.
'lm' for linear regression, 'glm' for generalized linear fit, 'loess' for local smooths.}

\item{se}{If 'TRUE', the confidence interval of the best-fit line will be shown.
If 'FALSE', the confidence interval will not be shown.
Check `geom_smooth` for more information.}

\item{line_size}{The thickness of the line. It can be set by giving a number.}

\item{separate_by}{This marks how the p- and r- values should be separated.
The default option is: ','
For more information, check out stat_cor() from the ggpubr package.}

\item{label_x}{Location of the statistical value prints along the figure's x-axis.
It asks for a number within the x-axis limit.}

\item{label_y}{Location of the statistical value prints along the figure's y-axis.
It requires a number within the y-axis limit.}

\item{text_size}{Size (numerical value) of the texts from correlation.}
}
\description{
This combines two different functions:
1) `geom_smooth()` from ggplot, and 2) `stat_cor()` from ggpubr.
`geom_smooth()` is used to fit the best-fit model, whereas
`stat_cor()` is used to print correlation results at an optimized location.
}
