% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_violin.R
\name{ses_violin}
\alias{ses_violin}
\title{violin plot with individual points (jittered)}
\usage{
ses_violin(
  violin_fill_color = "gray90",
  violin_border_color = "transparent",
  points = TRUE,
  point_size = 2.5,
  point_shape = 21,
  point_border_color = "white",
  ...
)
}
\arguments{
\item{violin_fill_color}{Color of the violin plot}

\item{violin_border_color}{Color of the violin's border.}

\item{points}{TRUE if points need to be shown.
FALSE if points need to be hidden.}

\item{point_size}{Size of the individual jittered points.}

\item{point_shape}{Shape of the jittered points.
Only shapes (21-25) with borders are allowed.}

\item{point_border_color}{Color of the points' border.}

\item{...}{Other parameters for geom_point(), such as "fill".
For more information check out ?geom_point.}
}
\description{
#' @description
A violin plot superimposed by jittered individual points.
}
\examples{
library(tidyverse)
library(sesplot)
# generate random data first
set.seed(1)
day1 = rnorm(20,0,1)
day2 = rnorm(20,5,1)
Subject <- rep(paste0('S',seq(1:20)), 2)
Data <- data.frame(Value = matrix(c(day1,day2),ncol=1))
Time <- rep(c('Day 1', 'Day 2'), each = length(day1))
df <- cbind(Subject, Data, Time)

# boxplot with all black points
ggplot(data = df, mapping = aes(x = Time, y = Value)) +
ses_violin(point_fill_color = 'black')

# boxplot with different colors of points across Time
ggplot(data = df, mapping = aes(x = Time, y = Value,
fill = Time)) + ses_violin()

}
