% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_auc_list.R
\name{sm_auc_list}
\alias{sm_auc_list}
\title{Calculating Area under Curve across multiple conditions and subjects}
\usage{
sm_auc_list(...)
}
\arguments{
\item{...}{Parameters of `sm_auc_all`}
}
\description{
This function is deprecated. Use `sm_auc_all` instead.
}
