% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_auc_list.R
\name{sm_auc_list}
\alias{sm_auc_list}
\title{Calculating Area under Curve across multiple conditions and subjects}
\usage{
sm_auc_list(subjects, conditions, x, values)
}
\arguments{
\item{subjects}{A column of 'subjects' from a data frame or a unique list of subjects}

\item{conditions}{A column of 'Conditions' from a data frame or a unique list of subjects}

\item{x}{A column of 'x' from a data frame or a unique list of x coordinate.}

\item{values}{A column of 'values' from a data frame that will be used to
the compute area under curve.}
}
\value{

}
\description{
This function returns a dataframe of AUC computed from raw data
('values' argument).
}
