% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_boxplot.R
\name{sm_boxplot}
\alias{sm_boxplot}
\title{A boxplot with individual points (jittered)}
\usage{
sm_boxplot(
  boxplot_fill_color = "gray95",
  boxplot_border_color = "black",
  points = TRUE,
  notch = F,
  width = 0.5,
  point_size = 2.5,
  outlier_label = FALSE,
  borders = TRUE,
  legends = FALSE,
  ...
)
}
\arguments{
\item{boxplot_fill_color}{Color of the boxplot.}

\item{boxplot_border_color}{Color of the boxplot's border.}

\item{points}{#' TRUE if points need to be shown.
FALSE if points need to be hidden.}

\item{notch}{TRUE to notch the shape of the boxplot.
FALSE to keep the boxplot's shape.}

\item{width}{Width of the boxplot.}

\item{point_size}{Size of the individual jittered points.}

\item{outlier_label}{Set to FALSE as default. If TRUE, another point that denotes the outlier
(as shown in the jittered points) will appear. In the TRUE condition,
two points that refer to the same individual point will appear
(one jittered, one outlier label).}

\item{borders}{If the border needs to be displayed, the input should be TRUE.
If the border is not needed, the input should be FALSE.}

\item{legends}{If the legend needs to be displayed, the input should be TRUE.
If the legend is not needed, the input should be FALSE.}

\item{...}{Other parameters for geom_point(), such as "alpha", fill", "color" and "shape".
For more information for the parameters, please check out ?geom_point.}
}
\description{
A boxplot superimposed by jittered individual points.
}
\examples{
\dontrun{
set.seed(1) # generate random data
day1 = rnorm(20,0,1)
day2 = rnorm(20,5,1)
Subject <- rep(paste0('S',seq(1:20)), 2)
Data <- data.frame(Value = matrix(c(day1,day2),ncol=1))
Day <- rep(c('Day 1', 'Day 2'), each = length(day1))
df <- cbind(Subject, Data, Day) # final dataframe

#use the dataframe to generate a boxplot

ggplot(data = df, mapping = aes(x = Day, y = Value, fill = Day)) +
sm_boxplot(shape = 21, color = 'white') +
scale_fill_manual(values = sm_color('blue','orange'))
}

}
