% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_common_legend.R
\name{sm_common_legend}
\alias{sm_common_legend}
\title{Creating a common legend for subplots}
\usage{
sm_common_legend(
  ...,
  x = 0.5,
  y = 0.5,
  title = FALSE,
  direction = "vertical",
  border = TRUE,
  legend_spacing = 0.5,
  border_color = "black"
)
}
\arguments{
\item{...}{Additional arguments to adjust the legend within the theme() function.
In usual cases, this argument should not be filled.}

\item{x}{Location of the legend along the x-axis. Default is the middle origin (0.5).}

\item{y}{Location of the legend along the y-axis. Default is the middle origin (0.5).}

\item{title}{Title of the legend. Input should be string}

\item{direction}{Direction of the legend. 'Horizontal' or 'Vertical'.}

\item{border}{If set TRUE, border around the legend will be created.
If set FALSE, the border will be removed.}

\item{legend_spacing}{Spacing within the legend.}

\item{border_color}{Color of the legend border}
}
\value{
It prints a legend on a blank plot. It can be used to create
a common legend for subplots.
}
\description{
Creating a common legend for subplots
}
\examples{
\dontrun{
df2 <- read_csv('https://www.smin95.com/amblyopia_random.csv')
df2 \%>\% select(Subject, SF, absBP, Condition) \%>\%
ggplot(aes(x = SF, y = absBP, shape = Condition, group = Condition)) +
 geom_point(color='black', fill = 'black', size = 2.5) +
 sm_common_legend(x = .5, y = 0.5 , title=FALSE) +
 scale_shape_manual(values = c(21,22),
                    labels = c('Condition 1 ',
                               'Condition 2 '))
}
}
