% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_corr_theme.R
\name{sm_corr_theme}
\alias{sm_corr_theme}
\title{SM theme for correlation plots}
\usage{
sm_corr_theme(legends = TRUE, borders = TRUE)
}
\arguments{
\item{legends}{If the legend needs to be displayed, the input should be TRUE.
If the legend is not needed, the input should be FALSE.}

\item{borders}{If the border needs to be displayed, the input should be TRUE.
If the border is not needed, the input should be FALSE.}
}
\description{
This theme has major vertical and horizontal grids. This is useful for plotting correlation.
sm_corr_theme() is exactly the same as sm_hvgrid().
}
