% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_hvgrid_minor.R
\name{sm_hvgrid_minor}
\alias{sm_hvgrid_minor}
\title{A theme with horizontal and vertical major and minor grids}
\usage{
sm_hvgrid_minor(legends = TRUE, borders = TRUE)
}
\arguments{
\item{legends}{If the legend needs to be displayed, the input should be TRUE.
If the legend is not needed, the input should be FALSE.}

\item{borders}{If the border needs to be displayed, the input should be `TRUE`.
If the border is not needed, the input should be `FALSE`.}
}
\value{
Returns a background theme that has both horizontal and vertical major and minor grids (ggplot2 output).
}
\description{
This theme has vertical and horizontal grids.
}
\examples{
library(ggplot2)
library(smplot2)
ggplot(data = mpg) +
geom_point(mapping = aes(x = displ, y = hwy, color = class)) +
sm_hvgrid_minor()

}
