% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_palette.R
\name{sm_palette}
\alias{sm_palette}
\title{A custom palette of colors}
\usage{
sm_palette(colorNum = 10)
}
\arguments{
\item{colorNum}{Number of colors (1-20).}
}
\value{
Returns a hex code in string vector. The input determines the length of the output.
}
\description{
This is a custom color palette of smplot2.
It returns up to 20 different colors with a high contrast.
}
\examples{
library(smplot2)
#' sm_palette(3) # returns 3 colors

}
