% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_plot_clean.R
\name{sm_plot_clean}
\alias{sm_plot_clean}
\title{Remove xticklabels and yticklabels in selected panels for proper subplotting}
\usage{
sm_plot_clean(all_plots, ncol, nrow)
}
\arguments{
\item{all_plots}{all_plots should be list, which should contain all panels
that are to be combined into one figure.}

\item{ncol}{Number of columns in the combined plot}

\item{nrow}{Number of rows in the combined plot}
}
\description{
Remove xticklabels and yticklabels in selected panels for proper subplotting
}
\examples{
\dontrun{
all_plots1 <- sm_plot_clean(all_plots, ncol=3,nrow=2)
}
}
