% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_put_together.R
\name{sm_put_together}
\alias{sm_put_together}
\title{Combining figures together}
\usage{
sm_put_together(
  all_plots,
  title,
  xlabel,
  ylabel,
  legend,
  ncol,
  nrow,
  panel_scale = 0.9,
  wRatio = 1.1,
  hRatio = 1.1,
  margin = 1
)
}
\arguments{
\item{all_plots}{all_plots should be list, which should contain all panels
that are to be combined into one figure.}

\item{title}{Title layer that will determine the main title of the combined plot.
This is created using sm_common_title().}

\item{xlabel}{xlabel layer that will determine the label of the combined plot's x-axis.
This is created using sm_common_xlabel().}

\item{ylabel}{ylabel layer that will determine the label of the combined plot's y-axis.
This is created using sm_common_ylabel().}

\item{legend}{ggplot() layer that has legend}

\item{ncol}{Number of columns in the combined plot}

\item{nrow}{Number of rows in the combined plot}

\item{panel_scale}{Scale of the panel. Default is set to 0.9 to reduce empty space
within and around each panel. The user can set to a value from 0 to 1 to
see what happens to the spacing within each panel and between panels.}

\item{wRatio}{This adjusts the ratio of the width of the first column to those of other columns.
By default, it is set to be 1.1x wider than that of other columns. The 0.1 difference
is to compensate for the loss of space due to y-ticks in the first column.}

\item{hRatio}{This adjusts the ratio of the height of the last row to those of other rows
By default, it is set to be 1.1x taller than that of other columns. The 0.1 difference
is to compensate for the loss of space due to x-ticks in the last row.}

\item{margin}{Empty space between panels can be specified using this argument. The number can be
negative to reduce the space further (ex. -0.5).}
}
\value{
Returns a combined figure.
}
\description{
Combining figures together
}
\examples{
\dontrun{
title <- sm_common_title('My title')
xlabel <- sm_common_xlabel('My x-axis ')
ylabel <- sm_common_ylabel('My y-axis')
plots_tgd <- sm_put_together(indv_plots, title, xlabel,ylabel,
ncol=3,nrow=3)
}
}
