% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_raincloud.R
\name{sm_raincloud}
\alias{sm_raincloud}
\title{A raincloud plot - a combination of jittered points, boxplots
and violin plots.}
\usage{
sm_raincloud(
  data,
  x,
  y,
  group,
  which_side = "right",
  sep_level = 3,
  vertical = TRUE,
  jitter_width = 0.09,
  point_size = 3,
  line_color = "gray80",
  violin_alpha = 0.3,
  boxplot_alpha = 1,
  line_alpha = 0.6,
  ...
)
}
\arguments{
\item{data}{Data frame. Each column has to be a variable.
Each row has to be an observation. This data will get used as
argument in ggplot().}

\item{x}{Column variable of the data that defines the level of the x-axis.
This is used to define the aes() in ggplot().}

\item{y}{Column variable of the data that defines the level of the y-axis.
This is used to define the aes() in ggplot().}

\item{group}{Group argument for aesthetics. It has to be set for the variable
that defines each observation (ex. id, subject, patient, etc).
This is used to define the aes() in geom_line().
This argument is used to draw a line that pairs each observation
across different levels of the x-axis.

If this argument is empty, no line will be drawn.}

\item{which_side}{String argument to specify the side of the boxplots and violinplots.
The options are: 'right', 'left', and 'mixed'. 'mixed' only works
when there are 2 levels in the x-axis. Otherwise, it will return an error.}

\item{sep_level}{A numerical value that controls the level of the separation among
the boxplot, violin plot and the points. The value can be 0-3.
If it's 0, all of these are clustered together. If it's 3, they are all
separated. 1 and 2 are somewhere in the middle.}

\item{vertical}{The orientation of the plots. The default is set to TRUE.
If you want the horizontal orientation of the plot, please set this argument
as FALSE.}

\item{jitter_width}{A numerical value that determines the degree of the jitter for each point. If its 0,
all the points will have no jitter (aligned along the y-axis).}

\item{point_size}{Size of the points.}

\item{line_color}{Color of the line. This argument will be ignored
if group argument is missing.}

\item{violin_alpha}{Transparency of the violin (0 to 1).}

\item{boxplot_alpha}{Transparency of the boxplot (0 to 1).}

\item{line_alpha}{Transparency of the lines (0 to 1). This argument will be ignored
if group argument is missing.}

\item{...}{Parameters for the points' aesthetics, such as 'fill', 'shape', 'color',
and 'alpha'. For more information, please type ?geom_point}
}
\description{
This function visualizes a raincloud plot. The creation of this package
has been inspired by the R package called 'raincloudplots' by Jordy van
Langen (https://github.com/jorvlan/raincloudplots).

This function has been created to allow more customisation than the functions
in the raincloudplots package. Also, this function automatically sorts the data given the
condition that the x-axis factor levels have been sorted properly.
}
\references{
Allen M, Poggiali D, Whitaker K et al. Raincloud plots: a multi-platform tool for robust data visualization [version 2; peer review: 2 approved]. Wellcome Open Res 2021, 4:63. DOI: 10.12688/wellcomeopenres.15191.2
}
