% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_slope_list.R
\name{sm_slope_list}
\alias{sm_slope_list}
\title{Calculating the slope across multiple conditions and subjects}
\usage{
sm_slope_list(...)
}
\arguments{
\item{...}{This is the parameter for `sm_slope_all`.}
}
\description{
This function is deprecated. Use `sm_slope_all` instead.
}
