% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_vgrid.R
\name{sm_vgrid}
\alias{sm_vgrid}
\title{Minimalistic theme with vertical major grids}
\usage{
sm_vgrid(legends = TRUE, borders = TRUE)
}
\arguments{
\item{legends}{If the legend needs to be displayed, the input should be TRUE.
If the legend is not needed, the input should be FALSE.}

\item{borders}{If the border needs to be displayed, the input should be TRUE.
If the border is not needed, the input should be FALSE.}
}
\value{
Returns a background theme with major vertical grids (ggplot2 output).
}
\description{
This theme has major vertical grids.
}
\examples{
library(ggplot2)
library(smplot2)

ggplot(data = mpg) +
geom_point(mapping = aes(x = displ, y = hwy, color = class)) +
sm_vgrid()

}
